random edits

+++ src/views/objects/edit.blade.php
@@ -68,7 +68,13 @@
 
 @section('side')
 	<p>{{ Lang::get('avalon::messages.objects_edit_help', array('title'=>$object->title)) }}</p>
-	{{ Form::open(array('method'=>'delete', 'action'=>array('ObjectController@destroy', $object->id))) }}
-	<button type="submit" class="btn btn-mini">{{ Lang::get('avalon::messages.objects_destroy') }}</button>
-	{{ Form::close() }}
+
+	@if (!$dependencies)
+		{{ Form::open(array('method'=>'delete', 'action'=>array('ObjectController@destroy', $object->id))) }}
+		<button type="submit" class="btn btn-mini">{{ Lang::get('avalon::messages.objects_destroy') }}</button>
+		{{ Form::close() }}
+	@else
+		<p>{{ Lang::get('avalon::messages.objects_dependencies', array('dependencies', $dependencies)) }}</p>
+	@endif
+
 @endsection
